move GranularSyncAction to sync module
