+++ sync/src/commonMain/kotlin/org/dhis2/mobile/sync/model/GranularSyncAction.kt
@@ -1,4 +1,4 @@
-package org.dhis2.utils.granularsync
+package org.dhis2.mobile.sync.model
 
 sealed interface GranularSyncAction {
     data object DisplaySyncSuccess : GranularSyncAction
